Implemented ImportC (#116)
* ImportC ready

* Reverting to 72e1ff4

* Fixed example config and added version(importc) condition

* Linux vscode debug & wrap const(void*)* fix

* Variant fix

+++ examples/asteroids/source/asteroids.d
@@ -36,7 +36,7 @@ class Asteroids : GodotScript!Node3D {
             addAsteroid();
     }
 
-    @Method _process(float delta) {
+    @Method _process(double delta) {
         cameraTarget.position = player.position;
 
         foreach (ch; getChildren(false)) {
